backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
channels>=4.0
channels-redis>=4.0
django-cors-headers>=3.15
python-dotenv>=1.0
